worker/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymysql==1.1.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
torch==2.1.0
transformers==4.35.2
faiss-cpu==1.7.4
matplotlib==3.8.2
seaborn==0.13.1